OrderedSet is used on imagesets in multi_dataset_handling too.
I didn't see that before because apparently it only causes test
failures in xia2. This commit moves the minimal OrderedSet
implementation to that file and imports it in split_experiments
